MOD : Dao클래스 추가, 컨벤션 수정, 팩토리클래스 추가
