+++ src/main/java/lotto/dao/LottoDao.java
@@ -29,54 +29,6 @@ public class LottoDao {
         }
     }
 
-    public static void addWinningLotto(WinningLotto winningLotto, int lottoRound) {
-        try {
-            String query = "INSERT INTO winningLotto VALUES (?,?,?)";
-            PreparedStatement pstm = DBConnection.getConnection().prepareStatement(query);
-
-            pstm.setInt(1, lottoRound);
-            pstm.setString(2, winningLotto.toString());
-            pstm.setInt(3, winningLotto.getBonus());
-
-            pstm.executeUpdate();
-        } catch (SQLException e) {
-            System.out.println(e.getMessage());
-        }
-    }
-
-    public static void addResult(List<String> results, int lottoRound) {
-        try {
-            String query = "INSERT INTO results (result, round) VALUES (?,?)";
-            PreparedStatement pstm = DBConnection.getConnection().prepareStatement(query);
-
-            for (String result : results) {
-                pstm.setString(1, result);
-                pstm.setInt(2, lottoRound);
-                pstm.addBatch();
-                pstm.clearParameters();
-            }
-
-            pstm.executeBatch();
-        } catch (SQLException e) {
-            System.out.println(e.getMessage());
-        }
-    }
-
-    public static void addResultInfo(int lottoRound, double sum, String returnRate) {
-        try {
-            String query = "INSERT INTO resultInfo VALUES(?,?,?)";
-            PreparedStatement statement = DBConnection.getConnection().prepareStatement(query);
-
-            statement.setInt(1, lottoRound);
-            statement.setInt(2, (int) sum);
-            statement.setString(3, returnRate);
-
-            statement.executeUpdate();
-        } catch (SQLException e) {
-            System.out.println(e.getMessage());
-        }
-    }
-
     public static int offerMaxRound() {
         try {
             String query = "SELECT MAX(round) FROM userLotto";
@@ -115,90 +67,4 @@ public class LottoDao {
     }
 
 
-    public static List<String> offerResults(int lottoRound) {
-        String query = "SELECT result FROM results WHERE round = ?";
-        List<String> numbers = new ArrayList<>();
-
-        try {
-            PreparedStatement pstm = DBConnection.getConnection().prepareStatement(query);
-
-            pstm.setInt(1, lottoRound);
-            ResultSet rs = pstm.executeQuery();
-
-            while (rs.next()) {
-                numbers.add(rs.getString(1));
-            }
-        } catch (SQLException e) {
-            System.out.println(e.getMessage());
-        }
-
-        return numbers;
-    }
-
-    public static String offerWinningNumber(int lottoRound) {
-        try {
-            String query = "SELECT winningLotto FROM winningLotto WHERE round = ?";
-            PreparedStatement pstm = DBConnection.getConnection().prepareStatement(query);
-            pstm.setInt(1, lottoRound);
-            ResultSet rs = pstm.executeQuery();
-
-            if (!rs.next()) return null;
-
-            return rs.getString(1);
-
-        } catch (SQLException e) {
-            System.out.println(e.getMessage());
-        }
-
-        return null;
-    }
-
-    public static int offerBonusNumber(int lottoRound) {
-        try {
-            String query = "SELECT bonus FROM winningLotto WHERE round = ?";
-            PreparedStatement pstm = DBConnection.getConnection().prepareStatement(query);
-            pstm.setInt(1, lottoRound);
-            ResultSet rs = pstm.executeQuery();
-
-            if (!rs.next()) return 0;
-
-            return rs.getInt(1);
-        } catch (SQLException e) {
-            System.out.println(e.getMessage());
-        }
-
-        return 0;
-    }
-
-    public static int offerPrize(int lottoRound) {
-        try {
-            String query = "SELECT prize FROM resultInfo WHERE round = ?";
-            PreparedStatement statement = DBConnection.getConnection().prepareStatement(query);
-            statement.setInt(1, lottoRound);
-            ResultSet rs = statement.executeQuery();
-
-            if (!rs.next()) return 0;
-            return rs.getInt(1);
-        } catch (SQLException e) {
-            System.out.println(e.getMessage());
-        }
-
-        return 0;
-    }
-
-    public static double offerReturnRate(int lottoRound) {
-        try {
-            String query = "SELECT returnRate FROM resultInfo WHERE round = ?";
-            PreparedStatement statement = DBConnection.getConnection().prepareStatement(query);
-            statement.setInt(1, lottoRound);
-            ResultSet rs = statement.executeQuery();
-
-            if (!rs.next()) return 0;
-            return rs.getDouble(1);
-        } catch (SQLException e) {
-            System.out.println(e.getMessage());
-        }
-
-        return 0;
-    }
 }
