+++ src/main/java/lotto/domain/Parser.java
@@ -14,7 +14,7 @@ public class Parser {
             userLotto.add(Number.of(Integer.parseInt(number)));
         }
 
-        return new Lotto(userLotto);
+        return LottoFactory.createLotto(userLotto);
     }
 
     public static List<Lotto> parseLotto(String[] numbers) {
@@ -22,7 +22,7 @@ public class Parser {
 
         for (String number : numbers) {
             String[] oneNumbers = number.split(DELIMITER);
-            userLotto.add(new Lotto(oneNumbers));
+            userLotto.add(LottoFactory.createLotto(oneNumbers));
         }
 
         return userLotto;
@@ -30,9 +30,9 @@ public class Parser {
 
     public static String[] parseLottoStrings(String userLottoString) {
         String temp = userLottoString
-                .replace("[[" , "")
+                .replace("[[", "")
                 .replace("]]", "")
-                .replace(", ",",");
+                .replace(", ", ",");
         return temp.split("\\],\\[");
     }
 }
